+++ src/Sulu/Bundle/ContactBundle/Entity/ContactRepository.php
@@ -15,13 +15,16 @@ use Doctrine\ORM\NoResultException;
 use Doctrine\ORM\Query;
 use Doctrine\ORM\QueryBuilder;
 use Sulu\Component\Persistence\Repository\ORM\EntityRepository;
+use Sulu\Component\SmartContent\Orm\DataProviderRepositoryInterface;
+use Sulu\Component\SmartContent\Orm\DataProviderRepositoryTrait;
 
 /**
- * Repository for the Codes, implementing some additional functions
- * for querying objects.
+ * Repository for the contacts, implementing some additional functions for querying objects.
  */
-class ContactRepository extends EntityRepository
+class ContactRepository extends EntityRepository implements DataProviderRepositoryInterface
 {
+    use DataProviderRepositoryTrait;
+
     /**
      * find a contact by id.
      *
@@ -179,10 +182,10 @@ class ContactRepository extends EntityRepository
     /**
      * Searches Entities by where clauses, pagination and sorted.
      *
-     * @param int|null   $limit   Page size for Pagination
-     * @param int|null   $offset  Offset for Pagination
+     * @param int|null $limit Page size for Pagination
+     * @param int|null $offset Offset for Pagination
      * @param array|null $sorting Columns to sort
-     * @param array|null $where   Where clauses
+     * @param array|null $where Where clauses
      *
      * @return array Results
      */
@@ -227,8 +230,12 @@ class ContactRepository extends EntityRepository
      *
      * @return array
      */
-    public function findByAccountId($accountId, $excludeContactId = null, $arrayResult = true, $onlyFetchMainAccounts = true)
-    {
+    public function findByAccountId(
+        $accountId,
+        $excludeContactId = null,
+        $arrayResult = true,
+        $onlyFetchMainAccounts = true
+    ) {
         $qb = $this->createQueryBuilder('c');
 
         // only fetch main accounts
@@ -258,8 +265,8 @@ class ContactRepository extends EntityRepository
      * Add sorting to querybuilder.
      *
      * @param QueryBuilder $qb
-     * @param array        $sorting
-     * @param string       $prefix
+     * @param array $sorting
+     * @param string $prefix
      *
      * @return QueryBuilder
      */
@@ -277,8 +284,8 @@ class ContactRepository extends EntityRepository
      * add pagination to querybuilder.
      *
      * @param QueryBuilder $qb
-     * @param int|null     $limit  Page size for Pagination
-     * @param int|null     $offset Offset for Pagination
+     * @param int|null $limit Page size for Pagination
+     * @param int|null $offset Offset for Pagination
      *
      * @return QueryBuilder
      */
@@ -295,8 +302,8 @@ class ContactRepository extends EntityRepository
      * add where to querybuilder.
      *
      * @param QueryBuilder $qb
-     * @param array        $where
-     * @param string       $prefix
+     * @param array $where
+     * @param string $prefix
      *
      * @return QueryBuilder
      */
@@ -391,5 +398,34 @@ class ContactRepository extends EntityRepository
         } catch (NoResultException $nre) {
             return;
         }
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function appendJoins(QueryBuilder $queryBuilder)
+    {
+        $queryBuilder->addSelect('emails')
+            ->addSelect('emailType')
+            ->addSelect('phones')
+            ->addSelect('phoneType')
+            ->addSelect('faxes')
+            ->addSelect('faxType')
+            ->addSelect('urls')
+            ->addSelect('urlType')
+            ->addSelect('tags')
+            ->addSelect('categories')
+            ->addSelect('translations')
+            ->leftJoin('entity.emails', 'emails')
+            ->leftJoin('emails.emailType', 'emailType')
+            ->leftJoin('entity.phones', 'phones')
+            ->leftJoin('phones.phoneType', 'phoneType')
+            ->leftJoin('entity.faxes', 'faxes')
+            ->leftJoin('faxes.faxType', 'faxType')
+            ->leftJoin('entity.urls', 'urls')
+            ->leftJoin('urls.urlType', 'urlType')
+            ->leftJoin('entity.tags', 'tags')
+            ->leftJoin('entity.categories', 'categories')
+            ->leftJoin('categories.translations', 'translations');
     }
 }