+++ src/Sulu/Component/SmartContent/ContentType.php
@@ -260,7 +260,7 @@ class ContentType extends ComplexContentType
             $pageSize = intval($params['max_per_page']->getValue());
 
             // resolve paginated filters
-            $data = $provider->resolveFilters(
+            $data = $provider->resolveResourceItems(
                 $filters,
                 $params,
                 $options,
@@ -269,7 +269,7 @@ class ContentType extends ComplexContentType
                 $pageSize
             );
         } else {
-            $data = $provider->resolveFilters($filters, $params, $options, (!empty($limit) ? intval($limit) : null));
+            $data = $provider->resolveResourceItems($filters, $params, $options, (!empty($limit) ? intval($limit) : null));
         }
 
         // append view data